+++ SerialRx/SerialRx/blob.h
@@ -8,19 +8,19 @@
 #define BLOBS_IN_LIST 10
 
 typedef struct blob_struct {
-	int16_t xLeft;
-	int16_t xRight;
-	int16_t yUpper;
-	int16_t yLower;
-	uint16_t numSamples;
+	int8_t xLeft;
+	int8_t xRight;
+	int8_t yUpper;
+	int8_t yLower;
+	uint8_t numSamples;
 } blobStruct_t;
 
 
 void ClearBlobs();
-void ClearBlobNumber(uint16_t i);
+void ClearBlobNumber(uint8_t i);
 void ClearFinalLineData();
-int16_t addPointToBlobList(int16_t x, int16_t y);
-void MergeSecondBlobIntoFirst(uint16_t firstBlob, uint16_t secondBlob);
+int8_t addPointToBlobList(int8_t x, int8_t y);
+void MergeSecondBlobIntoFirst(uint8_t firstBlob, uint8_t secondBlob);
 uint16_t GetBlobCount();
 
 //} __attribute__((packed)) blobStruct_t;
